requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cors==1.0.1
cycler==0.10.0
DateTime==4.3
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
gevent==21.8.0
greenlet==1.1.1
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
patsy==0.5.1
Pillow==8.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
scipy==1.7.1
six==1.16.0
smartapi-python==1.2.9
statsmodels==0.12.2
tldextract==3.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
views==0.3
websocket-client==1.1.0
Werkzeug==2.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0